adding validation

+++ src/main/java/com/example/task_management_system/service/impl/TaskServiceImpl.java
@@ -3,18 +3,21 @@ package com.example.task_management_system.service.impl;
 import com.example.task_management_system.dto.TaskDto;
 import com.example.task_management_system.dto.TaskFilter;
 import com.example.task_management_system.dto.TaskResponse;
+import com.example.task_management_system.exceptions.AccessDeniedException;
 import com.example.task_management_system.exceptions.TaskNotFoundException;
 import com.example.task_management_system.exceptions.UserNotFoundException;
 import com.example.task_management_system.models.Task;
-import com.example.task_management_system.models.User;
+import com.example.task_management_system.models.UserEntity;
 import com.example.task_management_system.repository.TaskRepository;
 import com.example.task_management_system.repository.UserRepository;
 import com.example.task_management_system.service.TaskService;
+import com.example.task_management_system.specifications.TaskSpecifications;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.stream.Collectors;
@@ -32,16 +35,22 @@ public class TaskServiceImpl implements TaskService {
     }
 
     @Override
-    public TaskDto createTask(int authorId, int executorId, TaskDto taskDto) {
-        User author = userRepository.findById(authorId).orElseThrow(() -> new UserNotFoundException("Author with this ID can not be found"));
-        User executor = userRepository.findById(executorId).orElseThrow(() -> new UserNotFoundException("Executor with this ID can not be found"));
+    public TaskDto createTask(int executorId, TaskDto taskDto) {
+        String email = SecurityContextHolder.getContext().getAuthentication().getName();
+        UserEntity currentUser = userRepository.findByEmail(email).orElseThrow(() -> new UserNotFoundException("User not found"));
+
+        if (!currentUser.getRole().getName().equals("ADMIN")) {
+            throw new AccessDeniedException("Only admins can create tasks");
+        }
+
+        UserEntity executor = userRepository.findById(executorId).orElseThrow(() -> new UserNotFoundException("Executor with this ID can not be found"));
 
         Task task = new Task();
         task.setTitle(taskDto.getTitle());
         task.setDescription(taskDto.getDescription());
         task.setStatus(taskDto.getStatus());
         task.setPriority(taskDto.getPriority());
-        task.setAuthor(author);
+        task.setAuthor(currentUser);
         task.setExecutor(executor);
 
         Task newTask = taskRepository.save(task);
@@ -60,7 +69,7 @@ public class TaskServiceImpl implements TaskService {
     public TaskResponse getAllTasks(int pageNo, int pageSize, TaskFilter taskFilter) {
         Pageable pageable = PageRequest.of(pageNo, pageSize);
 
-        Specification<Task> spec = getSpecification(taskFilter);
+        Specification<Task> spec = TaskSpecifications.getSpecification(taskFilter);
 
         Page<Task> pageResult = taskRepository.findAll(spec, pageable);
 
@@ -84,6 +93,13 @@ public class TaskServiceImpl implements TaskService {
 
     @Override
     public TaskDto updateTask(int taskId, TaskDto taskDto) {
+        String email = SecurityContextHolder.getContext().getAuthentication().getName();
+        UserEntity currentUser = userRepository.findByEmail(email).orElseThrow(() -> new UserNotFoundException("User not found"));
+
+        if (!currentUser.getRole().getName().equals("ADMIN")) {
+            throw new AccessDeniedException("Only admins or executors can update tasks");
+        }
+
         Task task = taskRepository.findById(taskId).orElseThrow(() -> new TaskNotFoundException("Task with this ID can not be found"));
 
         task.setTitle(taskDto.getTitle());
@@ -98,6 +114,13 @@ public class TaskServiceImpl implements TaskService {
 
     @Override
     public void deleteTask(int taskId) {
+        String email = SecurityContextHolder.getContext().getAuthentication().getName();
+        UserEntity currentUser = userRepository.findByEmail(email).orElseThrow(() -> new UserNotFoundException("User not found"));
+
+        if (!currentUser.getRole().getName().equals("ADMIN")) {
+            throw new AccessDeniedException("Only admins can delete tasks");
+        }
+
         Task task = taskRepository.findById(taskId).orElseThrow(() -> new TaskNotFoundException("Task with this ID can not be found"));
         taskRepository.delete(task);
     }
@@ -111,51 +134,5 @@ public class TaskServiceImpl implements TaskService {
         taskDto.setPriority(task.getPriority());
 
         return taskDto;
-    }
-
-    private Specification<Task> getSpecification(TaskFilter filter) {
-        Specification<Task> spec = Specification.where(null);
-
-        if (filter.getTitle() != null) {
-            spec = spec.and(titleContains(filter.getTitle()));
-        }
-
-        if (filter.getStatus() != null) {
-            spec = spec.and(statusEquals(filter.getStatus()));
-        }
-
-        if (filter.getPriority() != null) {
-            spec = spec.and(priorityEquals(filter.getPriority()));
-        }
-
-        if (filter.getAuthorId() != null) {
-            spec = spec.and(authorEquals(Long.valueOf(filter.getAuthorId())));
-        }
-
-        if (filter.getExecutorId() != null) {
-            spec = spec.and(executorEquals(Long.valueOf(filter.getExecutorId())));
-        }
-
-        return spec;
-    }
-
-    private Specification<Task> titleContains(String title) {
-        return (root, query, criteriaBuilder) -> criteriaBuilder.like(root.get("title"), "%" + title + "%");
-    }
-
-    private Specification<Task> statusEquals(String status) {
-        return (root, query, criteriaBuilder) -> criteriaBuilder.equal(root.get("status"), status);
-    }
-
-    private Specification<Task> priorityEquals(String priority) {
-        return (root, query, criteriaBuilder) -> criteriaBuilder.equal(root.get("priority"), priority);
-    }
-
-    private Specification<Task> authorEquals(Long authorId) {
-        return (root, query, criteriaBuilder) -> criteriaBuilder.equal(root.join("author").get("id"), authorId);
-    }
-
-    private Specification<Task> executorEquals(Long executorId) {
-        return (root, query, criteriaBuilder) -> criteriaBuilder.equal(root.join("executor").get("id"), executorId);
     }
 }
